// src/app/features/unreal-pnl-total-query/models/search-params.model.ts
export interface SearchParams {
  APISERVER: string;
  sid: string;
  sip: string;
  Invscode: string;
  comp: string;
  cseq: string;
  stock: string;
  stockName: string;
  ctype: string;
  idno: string;
  ttype: string;
  exchangeRate: string;
  checkbox1: string;
  checkbox2: string;
  checkbox3: string;
}